# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agent Economy OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stripe/connect/onboard:
    post:
      operationId: startConnectOnboarding
      tags: [stripe]
      summary: Provision (or refresh) a Stripe Connect Express onboarding link
      responses:
        "200":
          description: Onboarding link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectOnboardingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /stripe/connect/status:
    get:
      operationId: getConnectStatus
      tags: [stripe]
      summary: Fresh Stripe Connect status for the current user
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Unauthorized — caller is not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ConnectOnboardingResponse:
      type: object
      properties:
        accountId:
          type: string
        onboardingUrl:
          type: string
        expiresAt:
          type: integer
        stub:
          type: boolean
      required: [accountId, onboardingUrl, expiresAt, stub]
    ConnectStatusResponse:
      type: object
      properties:
        accountId:
          type: ["string", "null"]
        status:
          type: string
          enum: [none, pending, verified, restricted]
        chargesEnabled:
          type: boolean
        payoutsEnabled:
          type: boolean
        requirementsCurrentlyDue:
          type: array
          items:
            type: string
        stub:
          type: boolean
      required: [status, chargesEnabled, payoutsEnabled, requirementsCurrentlyDue, stub]
...